revert: remove security.session_idle_timeout_minutes from OSS organization_settings (CAURA-660)
Reverts the OSS-side parts of #89 that put the per-org session idle
timeout setting in OSS organization_settings. Session timeout is a
UI-session feature (browser tabs, mouse activity) — not an OSS
concept, no agent or API auth involved — so it has no business
living on the OSS side. The follow-up CAURA-661/-662/-663/-665/-666
moves the whole feature to enterprise: per-org settings table on
platform-storage-api, platform-operations service for cleanup cron,
last_activity tracking on platform-auth-api, idle detector on the
enterprise frontend.

Reverted:
- common/security.py (deleted; the SESSION_IDLE_TIMEOUT_* constants
  move to the enterprise side)
- security block in DEFAULT_SETTINGS
- security.session_idle_timeout_minutes in _LEAF_TYPES + _LEAF_RANGES
- ResolvedConfig.session_idle_timeout_minutes getter
- The session-specific tests in test_session_idle_settings.py

Kept (generic bug fix unrelated to the architectural pivot):
- The bool/int validator guard in _validate_leaf_types.
  Renamed test_session_idle_settings.py to
  test_settings_validator_bool_int.py since the file's only purpose
  now is the generic guard's coverage. Two int fields remain that
  the guard protects: ``memory_retention_days`` and
  ``alert_critical_findings_min``.
